backend/requirements.txt
# InfoVault Backend Dependencies
# Flask and Core
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
pymongo==4.6.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2

# AWS S3
boto3==1.34.14
botocore==1.34.14

# Pinecone Vector Database
pinecone-client==3.0.0

# Hugging Face ML Models
transformers==4.38.2
sentence-transformers==2.2.2
torch==2.1.2
accelerate==0.27.2

# PDF Processing
pypdf==4.2.0

# Image OCR (optional - requires Tesseract installed)
pytesseract==0.3.10
Pillow==10.2.0

# HTTP Requests (for captcha verification)
requests==2.31.0

# Scheduling (for background jobs)
APScheduler==3.10.4

# Development
pytest==7.4.4
pytest-flask==1.3.0
